requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
ane_transformers==0.1.3
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
async-timeout==5.0.1
attrs==25.1.0
audioread==3.0.1
blinker==1.9.0
blobfile==3.0.0
cairocffi==1.7.1
CairoSVG==2.7.1
cattrs==24.1.2
certifi==2023.7.22
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
coremltools==8.2
cssselect2==0.8.0
cycler==0.12.1
datasets==3.4.1
decorator==5.2.1
defusedxml==0.7.1
diagrams==0.24.4
dill==0.3.8
discord.py==2.5.2
distlib==0.3.7
einops==0.8.1
exceptiongroup==1.1.3
filelock==3.12.3
Flask==3.1.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
graphviz==0.20.3
huggingface-hub==0.28.1
identify==2.6.9
idna==3.4
imageio==2.37.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
lxml==5.3.1
MarkupSafe==3.0.2
matplotlib==3.10.1
more-itertools==10.6.0
moshi==0.2.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.2.0
multiprocess==0.70.16
networkx==3.4.2
nodeenv==1.9.1
numba==0.61.0
numpy==1.26.4
omegaconf==2.3.0
openai-whisper==20240930
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pipenv==2023.9.8
platformdirs==3.10.0
pluggy==1.5.0
pooch==1.8.2
pre_commit==4.1.0
propcache==0.3.0
protobuf==3.20.1
psutil==7.0.0
pyaml==25.1.0
pyarrow==19.0.1
pycparser==2.22
pycryptodomex==3.22.0
pydantic==2.3.0
pydantic_core==2.6.3
pydub==0.25.1
pygame==2.6.1
pyparsing==3.2.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
sentencepiece==0.2.0
shapely==2.0.7
silentcipher @ git+https://github.com/SesameAILabs/silentcipher@d46d7d0893a583d8968ab3a6626e2289faec9152
six==1.17.0
sknw==0.15
sniffio==1.3.0
sounddevice==0.5.0
soundfile==0.13.1
soxr==0.5.0.post1
sphn==0.1.8
starlette==0.27.0
svgpathtools==1.6.1
svgwrite==1.4.3
sympy==1.13.3
threadpoolctl==3.6.0
tifffile==2025.3.13
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.0
tomli==2.2.1
torch==2.4.0
torchao==0.9.0
torchaudio==2.4.0
torchtune==0.4.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
virtualenv==20.24.5
webencodings==0.5.1
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.18.3
